# Add Cohere Command R7B, replace older Command R+ handler