--- a/berkeley-function-call-leaderboard/bfcl/model_handler/constant.py
+++ b/berkeley-function-call-leaderboard/bfcl/model_handler/constant.py
@@ -149,6 +149,7 @@ UNDERSCORE_TO_DOT = [
     "NousResearch/Hermes-2-Pro-Mistral-7B",
     "NousResearch/Hermes-2-Theta-Llama-3-8B",
     "NousResearch/Hermes-2-Theta-Llama-3-70B",
+    "command-r-plus-FC",
     "command-r7b-12-2024-FC",
     "THUDM/glm-4-9b-chat",
     "ibm-granite/granite-20b-functioncalling",