--- a/berkeley-function-call-leaderboard/pyproject.toml
+++ b/berkeley-function-call-leaderboard/pyproject.toml
@@ -25,7 +25,7 @@ dependencies = [
     "openai==1.46.0",
     "mistralai==1.1.0",
     "anthropic==0.39.0",
-    "cohere==5.5.8",
+    "cohere==5.13.3",
     "typer>=0.12.5",
     "tabulate>=0.9.0",
     "datamodel-code-generator==0.25.7",