Greetings! It's been a while since our last contribution to BFCL, and the new versions and recent improvements are great to see.

This PR adds our latest model, [released on Friday](https://cohere.com/blog/command-r7b). I've also replaced our older models because it simplifies the code within the BFCL framework.

When I run this PR against [3245d9](https://github.com/ShishirPatil/gorilla/commit/3245d9693bd93e5578920420daeb15fd60ee9eb0) I get the following results (without REST category sanity checks):

```
|Name |Overall Acc|Model      |Model Link                         |Cost ($ Per 1k Function Calls)|Latency Mean (s)|Latency Standard Deviation (s)|Latency 95th Percentile (s)|Non-Live AST Acc|Non-Live Simple AST|Non-Live Multiple AST|Non-Live Parallel AST|Non-Live Parallel Multiple AST|Non-Live Exec Acc|Non-Live Simple Exec|Non-Live Multiple Exec|Non-Live Parallel Exec|Non-Live Parallel Multiple Exec|Live Acc|Live Simple AST|Live Multiple AST|Live Parallel AST|Live Parallel Multiple AST|Multi Turn Acc|Multi Turn Base|Multi Turn Miss Func|Multi Turn Miss Param|Multi Turn Long Context|Relevance Detection|Irrelevance Detection|Organization|License     |
|-----|-----------|-----------|-----------------------------------|------------------------------|----------------|------------------------------|---------------------------|----------------|-------------------|---------------------|---------------------|------------------------------|-----------------|--------------------|----------------------|----------------------|-------------------------------|--------|---------------|-----------------|-----------------|--------------------------|--------------|---------------|--------------------|---------------------|-----------------------|-------------------|---------------------|------------|------------|
|Run 1|53.37%     |Command R7B|https://cohere.com/blog/command-r7b|0.1                           |5.22            |9.75                          |12.64                      |81.00%          |68.00%             |91.50%               |84.50%               |80.00%                        |81.21%           |84.86%              |90.00%                |80.00%                |70.00%                         |74.23%  |63.95%         |69.71%           |50.00%           |70.83%                    |5.12%         |7.00%          |1.00%               |6.50%                |6.00%                  |61.11%             |80.68%               |Cohere      |cc-by-nc-4.0|
|Run 2|53.59%     |Command R7B|https://cohere.com/blog/command-r7b|0.1                           |3.19            |10.43                         |6.29                       |81.21%          |68.33%             |91.50%               |84.50%               |80.50%                        |82.46%           |85.36%              |92.00%                |80.00%                |72.50%                         |74.28%  |63.18%         |69.71%           |50.00%           |83.33%                    |5.12%         |7.50%          |1.50%               |6.00%                |5.50%                  |61.11%             |80.47%               |Cohere      |cc-by-nc-4.0|
```

I notice some variance between runs but haven't concluded where this is coming from - I'm confident, however, that the model regularly scores the higher score.

Thanks in advance for reviewing this PR and I look forward to seeing us on the leaderboard!