@Fanjia-Yan 2024-12-16
Hi Cohere,

Congrats on the release and thanks for contributing the really nice tool calling interface to interact with command models! We have one question before onboarding r7b on the leaderboard:

We typically don't deprecate a model unless the next generation alternative releases. With that, would v2 client be able to support full tool usage capability of Command R+ and Command R now or in the future? We do like to see models of different sizes on leaderboard for better comparison :) Thank you.

cc: @HuanzhiMao @ShishirPatil