@harry-cohere 2024-12-17
Thanks @Fanjia-Yan ! I've added `command-r-plus` back and it now uses the updated handler. When I run it against [3245d9](https://github.com/ShishirPatil/gorilla/commit/3245d9693bd93e5578920420daeb15fd60ee9eb0) I get this score:

```
|Overall Acc|Model              |Model Link                                           |Cost ($ Per 1k Function Calls)|Latency Mean (s)|Latency Standard Deviation (s)|Latency 95th Percentile (s)|Non-Live AST Acc|Non-Live Simple AST|Non-Live Multiple AST|Non-Live Parallel AST|Non-Live Parallel Multiple AST|Non-Live Exec Acc|Non-Live Simple Exec|Non-Live Multiple Exec|Non-Live Parallel Exec|Non-Live Parallel Multiple Exec|Live Acc|Live Simple AST|Live Multiple AST|Live Parallel AST|Live Parallel Multiple AST|Multi Turn Acc|Multi Turn Base|Multi Turn Miss Func|Multi Turn Miss Param|Multi Turn Long Context|Relevance Detection|Irrelevance Detection|Organization |License     |
|-----------|-------------------|-----------------------------------------------------|------------------------------|----------------|------------------------------|---------------------------|----------------|-------------------|---------------------|---------------------|------------------------------|-----------------|--------------------|----------------------|----------------------|-------------------------------|--------|---------------|-----------------|-----------------|--------------------------|--------------|---------------|--------------------|---------------------|-----------------------|-------------------|---------------------|-------------|------------|
|51.17%     |Command-R-Plus (FC)|https://txt.cohere.com/command-r-plus-microsoft-azure|7.71                          |2.75            |9.26                          |4.49                       |76.56%          |72.25%             |89.50%               |82.00%               |62.50%                        |79.91%           |90.14%              |90.00%                |82.00%                |57.50%                         |63.26%  |73.64%         |71.79%           |43.75%           |58.33%                    |15.00%        |23.00%         |10.00%              |11.50%               |15.50%                 |94.44%             |50.57%               |Cohere For AI|cc-by-nc-4.0|
```

This is +5.1% from the leaderboard score, and the increase mainly comes from the multihop category.

Looking into the increase I think part of the reason is that there are bugs in the replaced handler which lowered the overall score. For example the base handler [adds a holdout function message with role/content](https://github.com/cohere-ai/internal-gorilla/blob/3245d9693bd93e5578920420daeb15fd60ee9eb0/berkeley-function-call-leaderboard/bfcl/model_handler/base_handler.py#L126), but [the replaced handler expects them as role/message](https://github.com/cohere-ai/internal-gorilla/blob/3245d9693bd93e5578920420daeb15fd60ee9eb0/berkeley-function-call-leaderboard/bfcl/model_handler/proprietary_model/cohere.py#L204-L211), which would explain the 10% increase in the missed function category.

Another change is the cost per 1K functions - I see an increase of around $2.30 per 1K functions. Is it possible that the price is increased because the multi turn categories now succeed in more cases?